Core/Auras: fixed a possible underflow when checking required powers for SPELL_AURA_TRIGGER_SPELL_ON_POWER_AMOUNT and SPELL_AURA_TRIGGER_SPELL_ON_POWER_PCT
Port From (https://github.com/TrinityCore/TrinityCore/commit/f26c94205336f9d8cdee4eb9b0aedb7847ea0746)

+++ Source/Game/Entities/StatSystem.cs
@@ -755,7 +755,7 @@ namespace Game.Entities
                 {
                     if (effect.GetMiscValue() == (int)power)
                     {
-                        uint effectAmount = (uint)effect.GetAmount();
+                        int effectAmount = effect.GetAmount();
                         uint triggerSpell = effect.GetSpellEffectInfo().TriggerSpell;
 
                         float oldValueCheck = oldVal;
